refactor: update folder structure, general app clean
