--- client/src/app/(Auth)/components/index.ts
@@ -1,1 +1,0 @@
-export { SignOut } from './SignOut';
