+++ client/src/app/notes/components/Note/index.styled.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import Link from 'next/link';
 import styled from 'styled-components';
