+++ client/src/app/notes/components/index.ts
@@ -1,6 +1,2 @@
-export { default as CreateNote } from './Actions/CreateNote';
-export { default as DeleteNote } from './Actions/DeleteNote';
-export { default as EditNote } from './Actions/EditNote';
-export { default as ToggleCompleteNote } from './Actions/ToggleCompleteNote';
 export { default as Note } from './Note';
 export { default as NotesContainer } from './NotesContainer';
