+++ client/src/components/Actions/DeleteNote/index.styled.ts
