+++ client/src/components/Actions/DeleteNote/types.ts
