+++ client/src/components/Actions/index.ts
@@ -1,0 +1,4 @@
+export { default as CreateNote } from './CreateNote';
+export { default as DeleteNote } from './DeleteNote';
+export { default as EditNote } from './EditNote';
+export { default as ToggleComplete } from './ToggleComplete';
