+++ client/src/components/Auth/SignOut/index.styled.ts
