+++ client/src/components/Auth/index.ts
@@ -1,0 +1,1 @@
+export { default as SignOut } from './SignOut';
