+++ client/src/components/index.ts
@@ -2,4 +2,4 @@
 export { default as Nav } from './Nav';
 
 // General UI Components
-export { Button, ButtonLink, Container, Typography } from './ui';
+export { Button, ButtonLink, Container, Typography } from './UI';
