+++ client/src/components/ui/Typography/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { CSSProperties } from 'react';
 
 import * as Styled from './index.styled';
